Squashed 'staging/operator-registry/' changes from 601581055..4ca51503c
4ca51503c Merge pull request #570 from rphillips/fixes/closes
8a08e4f05 Merge pull request #576 from openshift-bot/art-consistency-openshift-4.8-operator-registry
741330866 Merge pull request #546 from timflannagan/minor-doc-fixes
92096e254 Merge pull request #580 from bparees/typo
50c93e26f Fix typo in deprecate+truncate command help
0e24b2bc3 Merge pull request #579 from timflannagan/fix-e2e-tests
22f0ada87 .github: Use the pre-installed Podman dependency
580c36f7e docs/design: Remove all references to the bundle alpha opm commands
12de30ec4 docs/design: Minor documentation fixes
8f3b1a423 Updating operator-registry builder & base images to be consistent with ART Reconciling with https://github.com/openshift/ocp-build-data/tree/a74468426e8d0524b2fca64bbbcda77afef25b80/images/operator-registry.yml
eb9fff530 Merge pull request #558 from ecordell/fix-upstream-example
904c60069 Merge pull request #559 from bentito/opm_doc_add_v1.15.3_idx_add
ae3636684 Merge pull request #571 from njhale/feat/no-sort
fd38fe618 fix(index): add build id ordering to channel head selection
a21f34ea7 test(e2e): make index add regression test pluggable
b29b4d251 (e2e) regression check with existing catalogs
97c2f954d add missing closes
6183dbb35 Merge pull request #572 from exdx/fix/test-minikube-driver
f88bcb4a8 fix: use none driver without sudo in upstream minikube e2e tests
bf4f88de5 Various from PR comments
16422cd79 Merge pull request #552 from exdx/feat/e2e-registry
6e870ac79 feat: use secure local registry for e2e tests
064f85a9d Merge pull request #556 from ecordell/fix-prune
cb6ec851b Fix upstream example dockerfile
32f790222 remove stranded bundles when packages are pruned
c2fc80a7d ensure foreign keys are always enabled
8e6c13fab Merge pull request #548 from exdx/fix/kind-ci
405e2e948 fix: load required docker images for kind cluster in ci e2e test
e3313a02d Merge pull request #541 from multi-arch/master
5afbd8a91 Updating Dockerfile baseimages to match ocp-build-data config
116a5cc43 Fix windows build with golang 1.15
REVERT: 601581055 build and related files
REVERT: 0a5f4475b vendoring
REVERT: b7bc1a653 Merge commit '6cf37ecef0072a2f5351f10c7de4edcbb361c0fc' as 'staging/operator-registry'

git-subtree-dir: staging/operator-registry
git-subtree-split: 4ca51503c460a2775c58c5cd6faa4baf09e6cc35

+++ pkg/sqlite/load.go
@@ -67,7 +67,6 @@ func (s *sqlLoader) AddOperatorBundle(bundle *registry.Bundle) error {
 }
 
 func (s *sqlLoader) addOperatorBundle(tx *sql.Tx, bundle *registry.Bundle) error {
-	// addBundle, err := tx.Prepare("insert into operatorbundle(name, csv, bundle, bundlepath, version, skiprange, replaces, skips, annotations) values(?, ?, ?, ?, ?, ?, ?, ?, ?)")
 	addBundle, err := tx.Prepare("insert into operatorbundle(name, csv, bundle, bundlepath, version, skiprange, replaces, skips) values(?, ?, ?, ?, ?, ?, ?, ?)")
 	if err != nil {
 		return err
@@ -80,7 +79,6 @@ func (s *sqlLoader) addOperatorBundle(tx *sql.Tx, bundle *registry.Bundle) error
 	}
 	defer addImage.Close()
 
-	// csvName, bundleImage, csvBytes, bundleBytes, annotationBytes, err := bundle.Serialize()
 	csvName, bundleImage, csvBytes, bundleBytes, _, err := bundle.Serialize()
 	if err != nil {
 		return err
@@ -107,7 +105,6 @@ func (s *sqlLoader) addOperatorBundle(tx *sql.Tx, bundle *registry.Bundle) error
 		return err
 	}
 
-	// if _, err := addBundle.Exec(csvName, csvBytes, bundleBytes, bundleImage, version, skiprange, replaces, strings.Join(skips, ","), annotationBytes); err != nil {
 	if _, err := addBundle.Exec(csvName, csvBytes, bundleBytes, bundleImage, version, skiprange, replaces, strings.Join(skips, ",")); err != nil {
 		return err
 	}
@@ -265,6 +262,10 @@ func (s *sqlLoader) AddPackageChannels(manifest registry.PackageManifest) error 
 	defer func() {
 		tx.Rollback()
 	}()
+
+	if err := s.rmPackage(tx, manifest.PackageName); err != nil {
+		return err
+	}
 
 	if err := s.addPackageChannels(tx, manifest); err != nil {
 		return err
@@ -439,7 +440,7 @@ func (s *sqlLoader) addPackageChannels(tx *sql.Tx, manifest registry.PackageMani
 				break
 			}
 			if _, _, _, err := s.getBundleSkipsReplacesVersion(tx, replaces); err != nil {
-				errs = append(errs, fmt.Errorf("%s specifies replacement that couldn't be found", c.CurrentCSVName))
+				errs = append(errs, fmt.Errorf("Invalid bundle %s, replaces nonexistent bundle %s", c.CurrentCSVName, replaces))
 				break
 			}
 
@@ -495,6 +496,7 @@ func (s *sqlLoader) getBundleSkipsReplacesVersion(tx *sql.Tx, bundleName string)
 		err = rerr
 		return
 	}
+	defer rows.Close()
 	if !rows.Next() {
 		err = fmt.Errorf("no bundle found for bundlename %s", bundleName)
 		return
@@ -535,6 +537,7 @@ func (s *sqlLoader) getBundlePathIfExists(tx *sql.Tx, bundleName string) (bundle
 		err = rerr
 		return
 	}
+	defer rows.Close()
 	if !rows.Next() {
 		// no bundlepath set
 		return
@@ -633,6 +636,9 @@ func (s *sqlLoader) getCSVNames(tx *sql.Tx, packageName string) ([]string, error
 	for rows.Next() {
 		err := rows.Scan(&csvName)
 		if err != nil {
+			if nerr := rows.Close(); nerr != nil {
+				return nil, nerr
+			}
 			return nil, err
 		}
 		csvNames = append(csvNames, csvName)
@@ -646,45 +652,51 @@ func (s *sqlLoader) getCSVNames(tx *sql.Tx, packageName string) ([]string, error
 }
 
 func (s *sqlLoader) RemovePackage(packageName string) error {
-	tx, err := s.db.Begin()
-	if err != nil {
-		return err
-	}
-	defer func() {
-		tx.Rollback()
-	}()
-
-	csvNames, err := s.getCSVNames(tx, packageName)
-	if err != nil {
-		return err
-	}
-	for _, csvName := range csvNames {
-		if err := s.rmBundle(tx, csvName); err != nil {
+	if err := func() error {
+		tx, err := s.db.Begin()
+		if err != nil {
 			return err
 		}
-	}
+		defer func() {
+			tx.Rollback()
+		}()
 
-	deletePackage, err := tx.Prepare("DELETE FROM package WHERE package.name=?")
-	if err != nil {
+		csvNames, err := s.getCSVNames(tx, packageName)
+		if err != nil {
+			return err
+		}
+		for _, csvName := range csvNames {
+			if err := s.rmBundle(tx, csvName); err != nil {
+				return err
+			}
+		}
+
+		deletePackage, err := tx.Prepare("DELETE FROM package WHERE package.name=?")
+		if err != nil {
+			return err
+		}
+		defer deletePackage.Close()
+
+		if _, err := deletePackage.Exec(packageName); err != nil {
+			return err
+		}
+
+		deleteChannel, err := tx.Prepare("DELETE FROM channel WHERE package_name = ?")
+		if err != nil {
+			return err
+		}
+		defer deleteChannel.Close()
+
+		if _, err := deleteChannel.Exec(packageName); err != nil {
+			return err
+		}
+		return tx.Commit()
+	}(); err != nil {
 		return err
 	}
-	defer deletePackage.Close()
 
-	if _, err := deletePackage.Exec(packageName); err != nil {
-		return err
-	}
-
-	deleteChannel, err := tx.Prepare("DELETE FROM channel WHERE package_name = ?")
-	if err != nil {
-		return err
-	}
-	defer deleteChannel.Close()
-
-	if _, err := deleteChannel.Exec(packageName); err != nil {
-		return err
-	}
-
-	return tx.Commit()
+	// separate transaction so that we remove stranded bundles after the package has been cleared
+	return s.RemoveStrandedBundles()
 }
 
 func (s *sqlLoader) rmBundle(tx *sql.Tx, csvName string) error {
@@ -758,32 +770,7 @@ func (s *sqlLoader) AddBundlePackageChannels(manifest registry.PackageManifest, 
 		return err
 	}
 
-	// Delete package, channel, and entries - they will be recalculated
-	deletePkg, err := tx.Prepare("DELETE FROM package WHERE name = ?")
-	if err != nil {
-		return err
-	}
-	defer deletePkg.Close()
-	_, err = deletePkg.Exec(manifest.PackageName)
-	if err != nil {
-		return err
-	}
-	deleteChan, err := tx.Prepare("DELETE FROM channel WHERE package_name = ?")
-	if err != nil {
-		return err
-	}
-	defer deleteChan.Close()
-	_, err = deleteChan.Exec(manifest.PackageName)
-	if err != nil {
-		return err
-	}
-	deleteChannelEntries, err := tx.Prepare("DELETE FROM channel_entry WHERE package_name = ?")
-	if err != nil {
-		return err
-	}
-	defer deleteChannelEntries.Close()
-	_, err = deleteChannelEntries.Exec(manifest.PackageName)
-	if err != nil {
+	if err := s.rmPackage(tx, manifest.PackageName); err != nil {
 		return err
 	}
 
@@ -792,6 +779,41 @@ func (s *sqlLoader) AddBundlePackageChannels(manifest registry.PackageManifest, 
 	}
 
 	return tx.Commit()
+}
+
+func (s *sqlLoader) rmPackage(tx *sql.Tx, pkg string) error {
+	// Delete package, channel, and entries - they will be recalculated
+	deletePkg, err := tx.Prepare("DELETE FROM package WHERE name = ?")
+	if err != nil {
+		return err
+	}
+
+	defer deletePkg.Close()
+	_, err = deletePkg.Exec(pkg)
+	if err != nil {
+		return err
+	}
+
+	deleteChan, err := tx.Prepare("DELETE FROM channel WHERE package_name = ?")
+	if err != nil {
+		return err
+	}
+
+	defer deleteChan.Close()
+	_, err = deleteChan.Exec(pkg)
+	if err != nil {
+		return err
+	}
+
+	deleteChannelEntries, err := tx.Prepare("DELETE FROM channel_entry WHERE package_name = ?")
+	if err != nil {
+		return err
+	}
+
+	defer deleteChannelEntries.Close()
+	_, err = deleteChannelEntries.Exec(pkg)
+
+	return err
 }
 
 func (s *sqlLoader) addDependencies(tx *sql.Tx, bundle *registry.Bundle) error {
@@ -958,6 +980,7 @@ func (s *sqlLoader) rmChannelEntry(tx *sql.Tx, csvName string) error {
 	}
 	for _, id := range entryIDs {
 		if _, err := updateChannelEntry.Exec(id); err != nil {
+			updateChannelEntry.Close()
 			return err
 		}
 	}
@@ -997,10 +1020,15 @@ func getTailFromBundle(tx *sql.Tx, name string) (bundles []string, err error) {
 		var skips sql.NullString
 		if rows.Next() {
 			if err := rows.Scan(&replaces, &skips); err != nil {
+				if nerr := rows.Close(); nerr != nil {
+					return nil, nerr
+				}
 				return nil, err
 			}
 		}
-		rows.Close()
+		if err := rows.Close(); err != nil {
+			return nil, err
+		}
 		if skips.Valid && skips.String != "" {
 			for _, skip := range strings.Split(skips.String, ",") {
 				tail[skip] = struct{}{}
@@ -1018,11 +1046,20 @@ func getTailFromBundle(tx *sql.Tx, name string) (bundles []string, err error) {
 				var defaultChannelHead sql.NullString
 				err := rows.Scan(&defaultChannelHead)
 				if err != nil {
+					if nerr := rows.Close(); nerr != nil {
+						return nil, nerr
+					}
 					return nil, err
 				}
 				if defaultChannelHead.Valid {
+					if nerr := rows.Close(); nerr != nil {
+						return nil, nerr
+					}
 					return nil, registry.ErrRemovingDefaultChannelDuringDeprecation
 				}
+			}
+			if err := rows.Close(); err != nil {
+				return nil, err
 			}
 			next = replaces.String
 			tail[replaces.String] = struct{}{}
@@ -1102,56 +1139,23 @@ func (s *sqlLoader) DeprecateBundle(path string) error {
 	return tx.Commit()
 }
 
-func (s *sqlLoader) RemoveStrandedBundles() ([]string, error) {
+func (s *sqlLoader) RemoveStrandedBundles() error {
 	tx, err := s.db.Begin()
 	if err != nil {
-		return nil, err
+		return err
 	}
 	defer func() {
 		tx.Rollback()
 	}()
 
-	bundles, err := s.rmStrandedBundles(tx)
-	if err != nil {
-		return nil, err
+	if err := s.rmStrandedBundles(tx); err != nil {
+		return err
 	}
 
-	return bundles, tx.Commit()
+	return tx.Commit()
 }
 
-func (s *sqlLoader) rmStrandedBundles(tx *sql.Tx) ([]string, error) {
-	strandedBundles := make([]string, 0)
-
-	strandedBundleQuery := `SELECT name FROM operatorbundle WHERE name NOT IN (select operatorbundle_name from channel_entry)`
-	rows, err := tx.QueryContext(context.TODO(), strandedBundleQuery)
-	if err != nil {
-		return nil, err
-	}
-
-	var name sql.NullString
-	for rows.Next() {
-		if err := rows.Scan(&name); err != nil {
-			return nil, err
-		}
-		if name.Valid {
-			strandedBundles = append(strandedBundles, fmt.Sprintf(`"%s"`, name.String))
-		}
-	}
-	rows.Close()
-
-	if len(strandedBundles) == 0 {
-		return nil, nil
-	}
-
-	rmStmt, err := tx.Prepare(fmt.Sprintf("DELETE FROM operatorbundle WHERE name IN(%s)", strings.Join(strandedBundles, ",")))
-	if err != nil {
-		return nil, err
-	}
-	defer rmStmt.Close()
-
-	if _, err := rmStmt.Exec(); err != nil {
-		return strandedBundles, err
-	}
-
-	return strandedBundles, nil
+func (s *sqlLoader) rmStrandedBundles(tx *sql.Tx) error {
+	_, err := tx.Exec("DELETE FROM operatorbundle WHERE name NOT IN(select operatorbundle_name from channel_entry)")
+	return err
 }
